Sprint 4: Business Logic and Access Control

+++ backend/src/main/java/com/planio/app/services/BoardService.java
@@ -3,6 +3,7 @@ package com.planio.app.services;
 import com.planio.app.dto.BoardDTO;
 import com.planio.app.entity.Board;
 import com.planio.app.entity.User;
+import com.planio.app.exceptions.ObjectNotFoundException;
 import com.planio.app.repositories.BoardRepository;
 import com.planio.app.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +12,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Stream;
 
 @Service
 @RequiredArgsConstructor
@@ -19,6 +21,8 @@ public class BoardService {
 
     private final BoardRepository boardRepository;
     private final UserRepository userRepository;
+    private final CurrentUserService currentUserService;
+    private final BoardAccessService boardAccessService;
 
     private BoardDTO mapToDTO(Board board) {
         BoardDTO boardDTO = new BoardDTO();
@@ -30,7 +34,8 @@ public class BoardService {
 
     public BoardDTO createBoard(BoardDTO boardDTO) {
         log.info("Creating board: {}", boardDTO.getName());
-        User owner = userRepository.findById(boardDTO.getOwnerId()).orElseThrow();
+        User owner = userRepository.findById(boardDTO.getOwnerId()).
+                orElseThrow(() -> new ObjectNotFoundException("User", boardDTO.getOwnerId()));
 
         Board board = Board.builder()
                 .name(boardDTO.getName())
@@ -47,44 +52,85 @@ public class BoardService {
 
     public void addParticipant(Long boardId, String email) {
         log.info("Adding {} to board {}", email, boardId);
-        Board board = boardRepository.findById(boardId).orElseThrow();
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("User not found"));
 
-        board.getParticipants().add(user);
+        User user = currentUserService.getCurrentUser();
+
+        Board board = boardRepository.findById(boardId)
+                .orElseThrow(() -> new ObjectNotFoundException("Board", boardId));
+
+        boardAccessService.checkAccess(board, user);
+
+        User participant = userRepository.findByEmail(email)
+                .orElseThrow(() -> new ObjectNotFoundException("User", email));
+
+        board.getParticipants().add(participant);
         boardRepository.save(board);
 
     }
 
     public List<BoardDTO> getAllBoards() {
         log.info("Getting all boards");
+
+        User user = currentUserService.getCurrentUser();
+
         return boardRepository.findAll()
                 .stream()
+                .filter(board -> boardAccessService.hasAccess(board, user))
                 .map(this::mapToDTO)
                 .toList();
     }
 
     public BoardDTO getBoardById(Long boardId) {
         log.info("Getting board with id: {}", boardId);
-        return  boardRepository.findById(boardId).map(this::mapToDTO)
-                .orElseThrow(() -> new RuntimeException("Board not found"));
+        Board board = boardRepository.findById(boardId)
+                .orElseThrow(() -> new ObjectNotFoundException("Board", boardId));
+
+        User user = currentUserService.getCurrentUser();
+
+        boardAccessService.checkAccess(board, user);
+
+        return mapToDTO(board);
     }
 
     public BoardDTO update(Long id, BoardDTO boardDTO) {
         log.info("Updating board id: {}", id);
+        User user = currentUserService.getCurrentUser();
 
-        Board board = boardRepository.findById(id).orElseThrow();
+        Board board = boardRepository.findById(id)
+                .orElseThrow(() -> new ObjectNotFoundException("Board", id));
+
+        boardAccessService.checkAccess(board, user);
+
         board.setName(boardDTO.getName());
-        Board updated = boardRepository.save(board);
 
-        log.info("Board updated id: {}", updated.getId());
+        log.info("Board updated id: {}", id);
 
-        return mapToDTO(updated);
+        return mapToDTO(boardRepository.save(board));
     }
 
     public void delete(Long id) {
         log.warn("Deleting board id: {}", id);
 
+        User user = currentUserService.getCurrentUser();
+
+        Board board = boardRepository.findById(id)
+                        .orElseThrow(() -> new ObjectNotFoundException("Board", id));
+
+        boardAccessService.checkAccess(board, user);
+
         boardRepository.deleteById(id);
+    }
+
+    public List<BoardDTO> getMyBoards() {
+
+        User user = currentUserService.getCurrentUser();
+
+        List<Board> owned = boardRepository.findByOwner(user);
+        List<Board> participant = boardRepository.findByParticipants_Id(user.getId());
+
+        return Stream.concat(owned.stream(), participant.stream())
+                .distinct()
+                .map(this::mapToDTO)
+                .toList();
     }
 }
